rename dsl bundle and change imports
