+++ tests/dev.arcovia.mitigation.sat.tests/src/dev/arcovia/mitigation/sat/dsl/tests/TUHHTestDSL.java
@@ -1,7 +1,7 @@
 package dev.arcovia.mitigation.sat.dsl.tests;
 
 import dev.arcovia.mitigation.sat.*;
-import dev.arcovia.mitigation.sat.dsl.CNFTranslation;
+import dev.arcovia.mitigation.utils.dsl.CNFTranslation;
 import dev.arcovia.mitigation.sat.dsl.tests.utility.CNFUtil;
 import dev.arcovia.mitigation.sat.tests.TUHHTest;
 import org.apache.log4j.Logger;
