+++ tests/dev.arcovia.mitigation.sat.tests/src/dev/arcovia/mitigation/sat/dsl/tests/unit/VariableTest.java
@@ -2,7 +2,7 @@ package dev.arcovia.mitigation.sat.dsl.tests.unit;
 
 import dev.arcovia.mitigation.sat.Constraint;
 import dev.arcovia.mitigation.sat.LabelCategory;
-import dev.arcovia.mitigation.sat.dsl.CNFTranslation;
+import dev.arcovia.mitigation.utils.dsl.CNFTranslation;
 import dev.arcovia.mitigation.sat.dsl.tests.utility.CNFUtil;
 import dev.arcovia.mitigation.sat.dsl.tests.dummy.DInData;
 import dev.arcovia.mitigation.sat.dsl.tests.dummy.DNode;
